feat: Queue타입 구현
